requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
ta==0.11.0
schedule==1.2.0
python-dotenv==1.0.0
pytz==2023.3